The "Nãome" fix.
Made so that during paramValue replacing it'll only replace whole words,
preventing things like "no" from being wrongfully replaced on words that
have "no" in them, such as "nome" (previously "name").

This is done by replacing an equal sign followed by the paramValue in question,
because if it's not a whole word, it won't have the equal sign in front of it.

A side effect from this change is that even if the english text had params and
equal signs separated by spaces, they will get removed in the translated version.