Refactored HTML and SASS structure for better redability.
Renamed most of the classes as to make sense to what they contain.

Removed one of the translator containers, as it was basically a dupe.

Removed the Header tag and file associated with it, as it was not being used.

Separated the containers in different lines in the __dropTable for better readability.

Translated a "search" to "pesquisar" since everything else is in portuguese.