Fixing the gemw, again.
Haste makes waste, as the saying goes.

I forgot to account for compound lines such as {{DropsLine}} where you
can't simply replace the whole thing like in a regular infobox (that has
only one paramName and Value pair per line, so I made it search for the gemw
exception with a regex to replace with the correct string instead of just
returning it directly.