Changing item jsons.
Changing back the dataItems.json to what it was before and giving the
search bar a dedicated json with only pt-br names and ids. Before, we
tried using the json from the Glouphrie Translator, which resulted in
some item IDs not being the correct version for the G.E. page.