Fixing the {{NPC map}} interaction.
The paramValues would break the whole code, and not only is this a
unique case due to the PT-BR version being untranslated (meaning, it's
still in English) but also the place where the exception happens differ
from the other exceptions, which makes the current way of handling
exceptions not work for this particular case.

As such, I just threw a big ol' if at the start to check for NPC map to
handle it whenever it arises, lol.